+++ src/test/java/nextstep/subway/station/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.station;
+package nextstep.subway.station.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
